+++ packages/google_maps_flutter/google_maps_flutter/lib/src/google_map.dart
@@ -594,7 +594,7 @@ class _GoogleMapState extends State<GoogleMap> {
 
 /// Builds a [MapConfiguration] from the given [map].
 MapConfiguration _configurationFromMapWidget(GoogleMap map) {
-  assert(!map.liteModeEnabled || Platform.isAndroid);
+  assert(!map.liteModeEnabled || isAndroid);
   return MapConfiguration(
     webGestureHandling: map.webGestureHandling,
     compassEnabled: map.compassEnabled,